Move inet_pton to posix platform-compatibility layer
