+++ src/posix.c
@@ -205,3 +205,5 @@ int p_write(git_file fd, const void *buf, size_t cnt)
 	}
 	return 0;
 }
+
+
